New project files and various deletions
