--- demo/graph-tutorial/example.env
@@ -1,4 +1,0 @@
-AZURE_APP_ID='YOUR_APP_ID_HERE'
-AZURE_CLIENT_SECRET='YOUR_CLIENT_SECRET_HERE'
-TLS_CERT_PATH='PATH_TO_LOCALHOST.CRT'
-TLS_KEY_PATH='PATH_TO_LOCALHOST.KEY'
