--- demo/graph-tutorial/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "graph-tutorial",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon ./src/server.ts",
-    "build": "tsc --project ./"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "@azure/msal-node": "1.3.0",
-    "@microsoft/microsoft-graph-client": "3.0.0",
-    "date-fns": "2.23.0",
-    "date-fns-tz": "1.1.6",
-    "dotenv": "10.0.0",
-    "express": "^4.17.1",
-    "express-promise-router": "^4.1.0",
-    "isomorphic-fetch": "^3.0.0",
-    "jsonwebtoken": "^8.5.1",
-    "jwks-rsa": "2.0.4",
-    "node-fetch": "^2.6.1",
-    "windows-iana": "5.0.2"
-  },
-  "devDependencies": {
-    "@types/express": "4.17.13",
-    "@types/isomorphic-fetch": "^0.0.35",
-    "@types/jquery": "3.5.6",
-    "@types/jsonwebtoken": "8.5.4",
-    "@types/microsoft-graph": "2.0.0",
-    "@types/node": "16.4.13",
-    "@types/node-fetch": "2.5.12",
-    "@types/office-js": "1.0.195",
-    "nodemon": "2.0.12",
-    "ts-node": "10.2.0",
-    "typescript": "4.3.5"
-  }
-}
