--- demo/graph-tutorial/src/addin/config.example.js
@@ -1,3 +1,0 @@
-authConfig = {
-  clientId: 'YOUR_APP_ID_HERE'
-};
