--- demo/graph-tutorial/src/addin/consent.html
@@ -1,17 +1,0 @@
-<!-- Copyright (c) Microsoft Corporation.
-     Licensed under the MIT License. -->
-
-<!-- <ConsentHtmlSnippet> -->
-<!DOCTYPE html>
-<html>
-  <head>
-    <script src="https://appsforoffice.microsoft.com/lib/beta/hosted/office.js"></script>
-    <script src="https://alcdn.msauth.net/browser/2.6.1/js/msal-browser.min.js"></script>
-    <script src="config.js"></script>
-    <script src="consent.js"></script>
-  </head>
-  <body class="ms-Fabric">
-    <p>Authenticating...</p>
-  </body>
-</html>
-<!-- </ConsentHtmlSnippet> -->
