--- demo/graph-tutorial/src/addin/taskpane.html
@@ -1,25 +1,0 @@
-<!-- Copyright (c) Microsoft Corporation.
-     Licensed under the MIT License. -->
-
-<!-- <TaskPaneHtmlSnippet> -->
-<html>
-  <head>
-    <link rel="stylesheet"
-      href="https://static2.sharepointonline.com/files/fabric/office-ui-fabric-core/11.0.0/css/fabric.min.css"/>
-    <link rel="stylesheet" href="taskpane.css"/>
-  </head>
-  <body class="ms-Fabric">
-    <div class="container">
-      <p class="ms-fontSize-32">Checking authentication...</p>
-    </div>
-    <div class="status"></div>
-    <div class="overlay">
-      <p class="ms-fontSize-24 ms-fontColor-white">Working...</p>
-    </div>
-    <script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.5.1.min.js"></script>
-    <script src="https://appsforoffice.microsoft.com/lib/beta/hosted/office.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/luxon@1.25.0/build/global/luxon.min.js"></script>
-    <script src="taskpane.js"></script>
-  </body>
-</html>
-<!-- </TaskPaneHtmlSnippet> -->
