+++ graph-tutorial/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "graph-tutorial",
+  "version": "1.0.0",
+  "description": "Office Spreadsheet addin using MS Graph",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "@azure/msal-node": "^1.13.0",
+    "@microsoft/microsoft-graph-client": "^3.0.2",
+    "date-fns": "^2.29.2",
+    "date-fns-tz": "^1.3.7",
+    "dotenv": "^16.0.2",
+    "express": "^4.18.1",
+    "express-promise-router": "^4.1.1",
+    "isomorphic-fetch": "^3.0.0",
+    "jsonwebtoken": "^8.5.1",
+    "jwks-rsa": "^2.1.4",
+    "node-fetch": "^3.2.10",
+    "windows-iana": "^5.1.0"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.13",
+    "@types/isomorphic-fetch": "^0.0.36",
+    "@types/jasmine": "^4.3.0",
+    "@types/jquery": "^3.5.14",
+    "@types/jsonwebtoken": "^8.5.9",
+    "@types/microsoft-graph": "^2.24.0",
+    "@types/node": "^18.7.16",
+    "@types/node-fetch": "^2.6.2",
+    "@types/office-js": "^1.0.276",
+    "jasmine-core": "^4.4.0",
+    "karma": "^6.4.0",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-jasmine": "^5.1.0",
+    "karma-webpack": "^5.0.0",
+    "nodemon": "^2.0.19",
+    "ts-loader": "^9.3.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.3",
+    "webpack": "^5.74.0",
+    "webpack-dev-server": "^4.11.0"
+  },
+  "scripts": {
+    "start": "nodemon ./src/server.ts",
+    "build": "tsc --project ./"
+  }
+}
