+++ graph-tutorial/src/addin/assets/icon-16.png
