+++ graph-tutorial/src/addin/assets/icon-32.png
