+++ graph-tutorial/src/addin/taskpane.css
@@ -1,0 +1,61 @@
+.container {
+    margin: 10px;
+  }
+  
+  .status {
+    margin: 10px;
+  }
+  
+  .overlay {
+    position: fixed;
+    display: none;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(0,0,0,0.5);
+    z-index: 2;
+    cursor: pointer;
+  }
+  
+  .overlay p {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%,-50%);
+  }
+  
+  .status-card {
+    padding: 1em;
+  }
+  
+  .primary-button {
+    padding: .5em;
+    color: white;
+    background-color: #0078d4;
+    border: none;
+    display: block;
+  }
+  
+  .success-msg {
+    background-color: #dff6dd;
+  }
+  
+  .error-msg {
+    background-color: #fde7e9;
+  }
+  
+  .date-picker {
+    display: block;
+    padding: .5em;
+    margin: 10px 0;
+  }
+  
+  .form-input {
+    display: block;
+    padding: .5em;
+    margin: 10px 0;
+    width: 100%;
+  }
