+++ graph-tutorial/src/addin/taskpane.js
@@ -1,0 +1,10 @@
+// TEMPORARY CODE TO VERIFY ADD-IN LOADS
+'use strict';
+
+Office.onReady(info => {
+  if (info.host === Office.HostType.Excel) {
+    $(function() {
+      $('p').text('Hello World!!');
+    });
+  }
+});
