+++ graph-tutorial/src/api/graph.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import Router from 'express-promise-router';
+
+const graphRouter = Router();
+
+// TODO: Implement this router
+
+export default graphRouter;
