--- tutorial/07-completed.md
@@ -1,7 +1,0 @@
-<!-- markdownlint-disable MD002 MD041 -->
-
-You've completed the Office Add-in Microsoft Graph tutorial. Now that you have a working add-in that calls Microsoft Graph, you can experiment and add new features. Visit the [Overview of Microsoft Graph](/graph/overview) to see all of the data you can access with Microsoft Graph.
-
-## Feedback
-
-Please provide any feedback on this tutorial in the [GitHub repository](https://github.com/microsoftgraph/msgraph-training-office-addin).
